changes in handling filenames for matching transf and csv files
Now it should handle more robustly inconsistent naming and still be able to find a match when available.
Now also printing in the command window the names of the files that have been matched, and which ones are being skipped.

A log is saved in the input_folder.

Corrected small issue